chore(next): restructures initPage
