+++ packages/next/src/exports/utilities.ts
@@ -5,4 +5,4 @@ export { createPayloadRequest } from '../utilities/createPayloadRequest.js'
 export { getNextRequestI18n } from '../utilities/getNextRequestI18n.js'
 export { getPayloadHMR, reload } from '../utilities/getPayloadHMR.js'
 export { headersWithCors } from '../utilities/headersWithCors.js'
-export { initPage } from '../utilities/initPage.js'
+export { initPage } from '../utilities/initPage/index.js'
